修改select不能默认选择的问题

+++ table.sql
@@ -68,7 +68,7 @@ CREATE TABLE sql_work
   is_backup TINYINT UNSIGNED NOT NULL DEFAULT 1 COMMENT '是否要备份，默认是备份',
   backup_table VARCHAR(50) NOT NULL DEFAULT '' COMMENT '备份之后的表名称',
   is_use_pt_osc TINYINT UNSIGNED NOT NULL DEFAULT 0 COMMENT '是否使用了pt-osc，默认没有使用',
-  sleep MEDIUMINT UNSIGNED NOT NULL DEFAULT 0 COMMENT '每条语句执行之后暂停多少毫秒，最小值0，最大值100秒，inception单位是毫秒，这边存储的是秒',
+  sleep FLOAT UNSIGNED NOT NULL DEFAULT 0 COMMENT '每条语句执行之后暂停多少毫秒，最小值0，最大值100秒，inception单位是毫秒，这边存储的是秒',
   sql_value MEDIUMTEXT COMMENT '要执行的sql内容',
   rollback_sql MEDIUMTEXT COMMENT '要回滚的SQL',
   audit_result_value MEDIUMTEXT COMMENT '审核的内容',
